+++ tests/ObjectTransformerTest.php
@@ -3,13 +3,16 @@ declare(strict_types=1);
 
 namespace Tests\LoyaltyCorp\Search;
 
-use LoyaltyCorp\Search\Transformer;
+use LoyaltyCorp\Search\Transformers\ObjectTransformer;
 use Tests\LoyaltyCorp\Search\Stubs\Handlers\Searches\NoDocumentBodyStub;
 use Tests\LoyaltyCorp\Search\Stubs\Handlers\Searches\NoSearchIdStub;
 use Tests\LoyaltyCorp\Search\Stubs\Handlers\Searches\SearchableStub;
 use Tests\LoyaltyCorp\Search\Stubs\Handlers\TransformableSearchHandlerStub;
 
-class TransformerTest extends TestCase
+/**
+ * @covers \LoyaltyCorp\Search\Transformers\ObjectTransformer
+ */
+class ObjectTransformerTest extends TestCase
 {
     /**
      * Tests the transformer correctly transforms a document.
@@ -19,7 +22,7 @@ class TransformerTest extends TestCase
     public function testTransform(): void
     {
         $handler = new TransformableSearchHandlerStub();
-        $transformer = new Transformer();
+        $transformer = new ObjectTransformer();
 
         $objects = [
             new NoSearchIdStub(),
